refactor: Set to use DbManager functions in AuthService - #103
